# api/openapi.yaml
openapi: 3.1.0
info:
  title: BlackPearl Local Setup API
  version: 0.1.0
  description: Localhost-only browser setup. TorBox tokens are write-only and never appear in responses.
servers:
  - url: http://localhost:8082
paths:
  /api/setup/status:
    get:
      operationId: getSetupStatus
      responses:
        "200":
          description: Public setup state and process-local CSRF value.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupStatus" }
  /api/setup/discover:
    post:
      operationId: discoverMedia
      description: First setup requires the launcher-provided bootstrap header and an explicit token. Reusing a saved token requires the setup-origin session header, bootstrap header, or exact token re-entry.
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/Session"
        - $ref: "#/components/parameters/Bootstrap"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                token: { type: string, minLength: 1, maxLength: 4096, writeOnly: true }
      responses:
        "200":
          description: Eligible completed account videos.
          headers:
            X-BlackPearl-Session:
              $ref: "#/components/headers/Session"
          content:
            application/json:
              schema:
                type: object
                required: [candidates]
                properties:
                  candidates: { type: array, items: { $ref: "#/components/schemas/MediaCandidate" } }
        "401": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/setup/configuration:
    put:
      operationId: applySetupConfiguration
      description: First setup requires the launcher-provided bootstrap header and an explicit token. Reconfiguration requires a setup-origin session, bootstrap header, or exact token re-entry.
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/Session"
        - $ref: "#/components/parameters/Bootstrap"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                token: { type: string, minLength: 1, maxLength: 4096, writeOnly: true }
                items:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { $ref: "#/components/schemas/ApplyItemRequest" }
                objectId: { type: string, pattern: "^[1-9][0-9]*:[1-9][0-9]*$" }
                title: { type: string, minLength: 1, maxLength: 200, description: Maximum 200 UTF-8 bytes. }
                year: { type: integer, minimum: 1888, maximum: 2100 }
      responses:
        "200":
          description: Activated public selection.
          headers:
            X-BlackPearl-Session:
              $ref: "#/components/headers/Session"
          content:
            application/json:
              schema:
                type: object
                required: [selected, selectedItems]
                properties:
                  selected: { $ref: "#/components/schemas/SetupConfiguration" }
                  selectedItems:
                    type: array
                    minItems: 1
                    maxItems: 100
                    items: { $ref: "#/components/schemas/SetupConfiguration" }
        "401": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/acquisition/status:
    get:
      operationId: getAcquisitionStatus
      description: Returns only whether private search-provider settings exist.
      responses:
        "200":
          description: Public non-secret acquisition configuration state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AcquisitionStatus" }
        "503": { $ref: "#/components/responses/Error" }
  /api/acquisition/settings:
    put:
      operationId: configureAcquisitionSearch
      description: Validates, probes, and privately stores one Prowlarr connection. Requires an already configured and paired TorBox setup.
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/Session"
        - $ref: "#/components/parameters/Bootstrap"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProwlarrSettingsInput" }
      responses:
        "200":
          description: Private settings were probed and saved.
          headers:
            X-BlackPearl-Session:
              $ref: "#/components/headers/Session"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AcquisitionStatus" }
        "401": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/acquisition/acquire:
    post:
      operationId: acquireMedia
      description: Searches configured indexers and publishes only a TorBox-cached movie or episode through the existing atomic Plex manifest transaction.
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/Session"
        - $ref: "#/components/parameters/Bootstrap"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcquisitionIntent" }
      responses:
        "200":
          description: Acquired media was published to the Plex manifest.
          headers:
            X-BlackPearl-Session:
              $ref: "#/components/headers/Session"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [selected, selectedItems]
                properties:
                  selected: { $ref: "#/components/schemas/SetupConfiguration" }
                  selectedItems:
                    type: array
                    minItems: 1
                    maxItems: 100
                    items: { $ref: "#/components/schemas/SetupConfiguration" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/watchlist/status:
    get:
      operationId: getWatchlistStatus
      description: Returns only privacy-safe aggregate observation state to an already paired local browser. No titles, identifiers, or credentials are returned.
      parameters:
        - $ref: "#/components/parameters/CSRF"
        - $ref: "#/components/parameters/Session"
        - $ref: "#/components/parameters/Bootstrap"
      responses:
        "200":
          description: Observe-only Plex watchlist health and queue counts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistStatus" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
components:
  parameters:
    CSRF:
      in: header
      name: X-BlackPearl-CSRF
      required: true
      schema: { type: string, minLength: 64, maxLength: 64 }
    Session:
      in: header
      name: X-BlackPearl-Session
      required: false
      description: Setup-origin browser authorization. It is never sent to Plex.
      schema: { type: string, pattern: "^[0-9a-f]{64}$" }
    Bootstrap:
      in: header
      name: X-BlackPearl-Bootstrap
      required: false
      description: Host-generated first-run pairing value carried in the setup URL fragment.
      schema: { type: string, pattern: "^[0-9a-f]{64}$" }
  headers:
    Session:
      required: true
      description: Setup-origin browser authorization derived from the saved TorBox token.
      schema: { type: string, pattern: "^[0-9a-f]{64}$" }
  schemas:
    MediaCandidate:
      type: object
      additionalProperties: false
      required: [objectId, name, extension, size]
      properties:
        objectId: { type: string }
        name: { type: string }
        extension: { type: string, enum: [.mp4, .mkv] }
        size: { type: integer, format: int64, minimum: 1 }
    SetupConfiguration:
      allOf:
        - $ref: "#/components/schemas/MediaCandidate"
        - type: object
          required: [mediaType, title, year]
          properties:
            mediaType: { type: string, enum: [movie, episode] }
            title: { type: string, maxLength: 200, description: Maximum 200 UTF-8 bytes. }
            year: { type: integer }
            showTitle: { type: string, maxLength: 200 }
            season: { type: integer, minimum: 0, maximum: 99 }
            episode: { type: integer, minimum: 1, maximum: 999 }
    ApplyItemRequest:
      type: object
      additionalProperties: false
      required: [objectId, title, year]
      properties:
        objectId: { type: string, pattern: "^[1-9][0-9]*:[1-9][0-9]*$" }
        mediaType: { type: string, enum: [movie, episode], default: movie }
        title: { type: string, minLength: 1, maxLength: 200, description: Maximum 200 UTF-8 bytes. }
        year: { type: integer, minimum: 1888, maximum: 2100 }
        showTitle: { type: string, minLength: 1, maxLength: 200 }
        season: { type: integer, minimum: 0, maximum: 99 }
        episode: { type: integer, minimum: 1, maximum: 999 }
    SetupStatus:
      type: object
      additionalProperties: false
      required: [setupRequired, tokenConfigured, csrfToken]
      properties:
        setupRequired: { type: boolean }
        tokenConfigured: { type: boolean }
        csrfToken: { type: string }
        selected: { $ref: "#/components/schemas/SetupConfiguration" }
        selectedItems:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/SetupConfiguration" }
    AcquisitionStatus:
      type: object
      additionalProperties: false
      required: [configured]
      properties:
        configured: { type: boolean }
    WatchlistStatus:
      type: object
      additionalProperties: false
      required: [enabled, healthy, queue]
      properties:
        enabled: { type: boolean }
        healthy: { type: boolean }
        lastSyncAt: { type: string, format: date-time }
        queue: { $ref: "#/components/schemas/WatchlistQueueStatus" }
    WatchlistQueueStatus:
      type: object
      additionalProperties: false
      required: [pendingMovies, acquiring, succeeded, notCached, retryable, manualReview, observedShows]
      properties:
        pendingMovies: { type: integer, minimum: 0 }
        acquiring: { type: integer, minimum: 0 }
        succeeded: { type: integer, minimum: 0 }
        notCached: { type: integer, minimum: 0 }
        retryable: { type: integer, minimum: 0 }
        manualReview: { type: integer, minimum: 0 }
        observedShows: { type: integer, minimum: 0 }
    ProwlarrSettingsInput:
      type: object
      additionalProperties: false
      required: [baseUrl, apiKey]
      properties:
        baseUrl:
          type: string
          minLength: 1
          maxLength: 2048
          description: Absolute HTTP(S) Prowlarr URL without credentials, query, or fragment.
        apiKey: { type: string, minLength: 1, maxLength: 4096, writeOnly: true }
    AcquisitionIntent:
      type: object
      additionalProperties: false
      required: [mediaType, title, year]
      properties:
        mediaType: { type: string, enum: [movie, episode] }
        title: { type: string, minLength: 1, maxLength: 200, description: Movie title or TV show title; maximum 200 UTF-8 bytes. }
        year: { type: integer, minimum: 1888, maximum: 2100 }
        season: { type: integer, minimum: 0, maximum: 99 }
        episode: { type: integer, minimum: 1, maximum: 999 }
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
  responses:
    Error:
      description: Public-safe error. It never includes tokens, signed URLs, or provider response bodies.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }